+++ digicert-certcentral-caplugin/Client/CertCentralClient.cs
@@ -314,6 +314,27 @@ namespace Keyfactor.Extensions.CAPlugin.DigiCert.Client
 			return orderResponse;
 		}
 
+		public OrderResponse OrderSmimeCertificate(OrderSmimeRequest request)
+		{
+			string jsonRequest = JsonConvert.SerializeObject(request, Formatting.None, new JsonSerializerSettings { NullValueHandling = NullValueHandling.Ignore });
+
+			Logger.LogTrace($"Order request:\n{jsonRequest}");
+
+			CertCentralResponse response = Request(request, jsonRequest);
+
+			OrderResponse orderResponse = new OrderResponse();
+			if (!response.Success)
+			{
+				Errors errors = JsonConvert.DeserializeObject<Errors>(response.Response);
+				orderResponse.Status = CertCentralBaseResponse.StatusType.ERROR;
+				orderResponse.Errors = errors.errors;
+			}
+			else
+				orderResponse = JsonConvert.DeserializeObject<OrderResponse>(response.Response);
+
+			return orderResponse;
+		}
+
 		public OrderResponse ReissueCertificate(ReissueRequest request)
 		{
 			string jsonRequest = JsonConvert.SerializeObject(request, Formatting.None, new JsonSerializerSettings { NullValueHandling = NullValueHandling.Ignore });